requirements.txt
easyocr==1.4.1
Flask==1.1.2
Flask_MySQLdb==1.0.1
imutils==0.5.4
MySQL-python==1.2.5
numpy==1.20.3
opencv_python_headless==4.5.5.62
pandas==1.3.4
Pillow==9.0.1
Pyrebase==3.0.27
python-dotenv==0.19.2
selenium==4.1.3
